refactor gemini generate image
